[new update]

+++ database/migrations/2022_11_11_073619_create_product_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->float('height');
             $table->float('length');
             $table->float('width');
-            $table->float('volume');
+            $table->double('volume');
             $table->date('import_date')->nullable();
             $table->date('export_date')->nullable();
             $table->float('weight')->nullable();
